+++ packages/dtslint/src/index.ts
@@ -5,12 +5,7 @@ import assert = require("assert");
 import fs from "fs";
 import { basename, dirname, join as joinPaths, resolve } from "path";
 
-import {
-  cleanTypeScriptInstalls,
-  deepEquals,
-  installAllTypeScriptVersions,
-  installTypeScriptNext,
-} from "@definitelytyped/utils";
+import { deepEquals } from "@definitelytyped/utils";
 import { checkPackageJson, checkTsconfig } from "./checks";
 import { checkTslintJson, lint, TsVersion } from "./lint";
 import { getCompilerOptions, packageNameFromPath } from "./util";
@@ -43,12 +38,6 @@ async function main(): Promise<void> {
       continue;
     }
     switch (arg) {
-      case "--installAll":
-        console.log("Cleaning old installs and installing for all TypeScript versions...");
-        console.log("Working...");
-        await cleanTypeScriptInstalls();
-        await installAllTypeScriptVersions();
-        return;
       case "--localTs":
         lookingForTsLocal = true;
         break;
@@ -88,26 +77,16 @@ async function main(): Promise<void> {
   }
 
   if (shouldListen) {
-    listen(dirPath, tsLocal, onlyTestTsNext);
+    listen(dirPath, tsLocal);
   } else {
-    await installTypeScriptAsNeeded(tsLocal, onlyTestTsNext);
     await runTests(dirPath, onlyTestTsNext, expectOnly, tsLocal);
   }
 }
 
-async function installTypeScriptAsNeeded(tsLocal: string | undefined, onlyTestTsNext: boolean): Promise<void> {
-  if (tsLocal) return;
-  if (onlyTestTsNext) {
-    return installTypeScriptNext();
-  }
-  return installAllTypeScriptVersions();
-}
-
 function usage(): void {
-  console.error("Usage: dtslint [--version] [--installAll] [--onlyTestTsNext] [--expectOnly] [--localTs path]");
+  console.error("Usage: dtslint [--version] [--onlyTestTsNext] [--expectOnly] [--localTs path]");
   console.error("Args:");
   console.error("  --version        Print version and exit.");
-  console.error("  --installAll     Cleans and installs all TypeScript versions.");
   console.error("  --expectOnly     Run only the ExpectType lint rule.");
   console.error("  --onlyTestTsNext Only run with `typescript@next`, not with the minimum version.");
   console.error("  --localTs path   Run with *path* as the latest version of TS.");
@@ -115,9 +94,8 @@ function usage(): void {
   console.error("onlyTestTsNext and localTs are (1) mutually exclusive and (2) test a single version of TS");
 }
 
-function listen(dirPath: string, tsLocal: string | undefined, alwaysOnlyTestTsNext: boolean): void {
+function listen(dirPath: string, tsLocal: string | undefined): void {
   // Don't await this here to ensure that messages sent during installation aren't dropped.
-  const installationPromise = installTypeScriptAsNeeded(tsLocal, alwaysOnlyTestTsNext);
   process.on("message", async (message: unknown) => {
     const { path, onlyTestTsNext, expectOnly } = message as {
       path: string;
@@ -125,7 +103,6 @@ function listen(dirPath: string, tsLocal: string | undefined, alwaysOnlyTestTsNe
       expectOnly?: boolean;
     };
 
-    await installationPromise;
     runTests(joinPaths(dirPath, path), onlyTestTsNext, !!expectOnly, tsLocal)
       .catch((e) => e.stack)
       .then((maybeError) => {